!2229 doc(#I4N6JR): update IDispatchService demo * doc: 更新消息分发示例
